Added base database interaction
Migration are automatically ran when the server gets started. Api requests now get stored in the database as a proof of concept. We might want to clean the API data since WCL sends prettified JSON sometimes. Tracking creation and access in case we need to setup some sort of auto-purge.